Example with division and casting to float

+++ Operators/Operators/Program.cs
@@ -13,7 +13,7 @@ namespace Operators
             var a = 10;
             var b = 3;
 
-            Console.WriteLine(a / b);
+            Console.WriteLine((float)a / (float)b);
         }
     }
 }
